Balancing two opponenents with your player on 2D table
On a 2D table positioned on an equilibrium point (0, 0) ‘others’ are having random walks within a designated area (red square ) and limited with a circular area to have the next move. And you try to bring the table to an equilibrium all the time with a move only limited with the table boundary. Simulate the requıred moves for 25 time instants.